requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pdfplumber>=0.10.0
python-multipart>=0.0.9
google-genai>=0.7.0
httpx>=0.27.0
groq>=0.9.0